refactor(git-hooks/): Extract code to `util.py`
